requirements.txt
python-dotenv==1.0.0
playwright==1.32.1
requests
